Refactor import paths to use discovery.config
