+++ packages/core/cypress/global.js
@@ -1,4 +1,4 @@
-import storeConfig from '../faststore.config'
+import storeConfig from '../discovery.config'
 
 export const { pages } = storeConfig.cypress
