+++ packages/core/discovery.config.default.js
