+++ packages/core/src/pages/account.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { NextSeo } from 'next-seo'
 
-import storeConfig from '../../faststore.config'
+import storeConfig from '../../discovery.config'
 import GlobalSections, {
   GlobalSectionsData,
   getGlobalSectionsData,
