+++ packages/core/src/server/cms/index.ts
@@ -3,7 +3,7 @@ import ClientCMS from '@vtex/client-cms'
 
 import MissingContentError from 'src/sdk/error/MissingContentError'
 import MultipleContentError from 'src/sdk/error/MultipleContentError'
-import config from '../../../faststore.config'
+import config from '../../../discovery.config'
 
 export type Options =
   | Locator
