+++ packages/core/src/server/cms/pdp.ts
@@ -3,7 +3,7 @@ import type { ContentData, Locator } from '@vtex/client-cms'
 import MissingContentError from 'src/sdk/error/MissingContentError'
 import { findBestPDPTemplate } from 'src/utils/multipleTemplates'
 import { Options, getCMSPage, getPage } from '.'
-import config from '../../../faststore.config'
+import config from '../../../discovery.config'
 
 type PDPSettings = {
   settings: {
